+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [v0.6.0] – 2016-06-14
 ### Added
 - `LitGroup\Sms\Exception\InvalidArgumentException`